# 4-Laravel/3-blog-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BLOG API Project",
        "description": "BLOG API Project",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
        "/api/comments/{post_id}": {
            "get": {
                "tags": [
                    "Comment"
                ],
                "summary": "get comments",
                "operationId": "GetComments",
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/comments": {
            "post": {
                "tags": [
                    "Comment"
                ],
                "summary": "store comment",
                "description": "Returns stored comment.",
                "operationId": "StoreComment",
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "comment_content",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/comments/approve/{comment_id}": {
            "get": {
                "tags": [
                    "Comment"
                ],
                "summary": "approve comment",
                "operationId": "approveComment",
                "parameters": [
                    {
                        "name": "comment_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/comments/unapprove/{comment_id}": {
            "get": {
                "tags": [
                    "Comment"
                ],
                "summary": "unapprove comment",
                "operationId": "unapproveComment",
                "parameters": [
                    {
                        "name": "comment_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "apiKey",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "name": "Authorization",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://127.0.0.1:8000/oauth/authorize",
                        "tokenUrl": "http://127.0.0.1:8000/oauth/token",
                        "refreshUrl": "http://127.0.0.1:8000/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}